Added settlement, road, and city placement scafold
I added some empty functions to handle road, settlement, and road placement
for the board. I have also decided that robber position can be denoted by
changing the resource value to a negative number. This will allow for the
resource's original identity to remain while also maintaining the position of
the robber.